Update to support new versions of moment.js
This is a proposed update for moment.js 2.7.0 + moment-timezone 0.1.0. Lines 24 to 28 were trying to preprocess the Olson IDs before they were added to the dataset in moment-timezone by using the "moment.tz.add" function. However, the new version of moment-timezone seems to have a get function that returns a sorted list of all available Olson IDs. I simply assigned the "names" array using this new function.